configs/rockpro64_ebbr: new defconfig
Add a defconfig to build an AArch64 disk image with a U-Boot based
firmware implementing the subset of UEFI defined by EBBR[1], as well as
a Linux OS booting with UEFI, to run on the ROCKPro64 from Pine64.

The generated firmware binary can also be used to install or run another
OS supporting the EBBR specification.

[1] https://github.com/ARM-software/ebbr

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>
Cc: Giulio Benetti <giulio.benetti@benettiengineering.com>
Cc: Jagan Teki <jagan@amarulasolutions.com>
Cc: Suniel Mahesh <sunil@amarulasolutions.com>
Signed-off-by: Julien Olivain <ju.o@free.fr>