package/qoriq-cadence-dp-firmware: bump version aligned to 6.12.3-1.0.0 BSP
This commit updates the qoriq-cadence-dp-firmware to the version
included in the 6.12.3-1.0.0 QorIQ BSP.

The firmware archive was refactored and its location has changed.
Those qoriq firmware files are now reused from the firmware-imx
archive, which is also included in Buildroot in the directory
"package/freescale-imx/firmware-imx". This commit keeps the Buildroot
package name qoriq-cadence-dp-firmware, first to keep compatibility
or Buildroot configs, and also because QorIQ BSP uses a different
version of the firmware-imx package, compared to the i.MX BSP (at a
same BSP version).

This commit updates the archive name, the archive version and the
download URL following the NXP Yocto BSP at [1]. More precisely,
the information was taken from the recipe at [2].

Note that regarding this qoriq-cadence-dp-firmware version:
- the QorIQ BSP version is: 6.12.3-1.0.0
- the Yocto dp-firmware-cadence package version is: 22.04
- the firmware-imx archive version is: 8.16

The important version we want to align with the rest of other QorIQ
components is indeed the BSP version "6.12.3-1.0.0".

This commit also updates license hashes. The license was updated from:
LA_OPT_NXP_Software_License v17 October 2020
to:
LA_OPT_NXP_Software_License v35 May 2022
Also, the license file EULA.txt was removed, so _LICENSE_FILES is
updated accordingly.

[1] https://github.com/nxp-qoriq/yocto-sdk
[2] https://github.com/Freescale/meta-freescale/blob/6fb8816cb32db4ba9391157036fa97914cc7d103/recipes-bsp/dp-firmware-cadence/dp-firmware-cadence_22.04.bb

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>
[Julien:
 - reword commit log to add info on different version schemes
 - fix license hashes
]
Signed-off-by: Julien Olivain <ju.o@free.fr>